Isolate unit tests with fixtures
We use Meck to mock Hackney, out HTTP client library of choice. But if
one of the tests fail, we never get to call `meck:unload`, and
subsequent `meck:new` calls always fail.

All the tests are now wrapped in a fixture that takes care of unloading
the module, regardless of test's outcome.