`create_pull_request`: required params are funargs
GitHub API is forced to pass some of the required parameters through
POST and not the URL because these parameters are too large. We aren't
constrained in this way, though, so we better make all required
parameters actual arguments to our functions.

This change strips us of a way to pass "body", but it will be brought
back by #36.